Refactored config to be a single point, this does mean sometimes we build things we don't need to
