+++ src/internal/config/config.go
@@ -1,4 +1,4 @@
-package main
+package config
 
 import (
 	"context"
@@ -6,9 +6,11 @@ import (
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	awsconfig "github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/service/lambda"
 	"github.com/aws/aws-xray-sdk-go/xray"
 	gogithub "github.com/google/go-github/v54/github"
 	"github.com/opentffoundation/registry/internal/github"
+	"github.com/opentffoundation/registry/internal/providers/providercache"
 	"github.com/opentffoundation/registry/internal/secrets"
 	"github.com/shurcooL/githubv4"
 	"os"
@@ -17,12 +19,18 @@ import (
 type Config struct {
 	ManagedGithubClient *gogithub.Client
 	RawGithubv4Client   *githubv4.Client
-	ProviderRedirects   map[string]string
 
-	SecretsHandler *secrets.Handler
+	LambdaClient         *lambda.Client
+	ProviderVersionCache *providercache.Handler
+	SecretsHandler       *secrets.Handler
+
+	ProviderRedirects map[string]string
 }
 
-func buildConfig(ctx context.Context) (config *Config, err error) {
+// BuildConfig will build a configuration object for the application. This
+// includes loading secrets from AWS Secrets Manager, and configuring the
+// AWS SDK.
+func BuildConfig(ctx context.Context, xraySegmentName string) (config *Config, err error) {
 	if err = xray.Configure(xray.Config{ServiceVersion: "1.2.3"}); err != nil {
 		err = fmt.Errorf("could not configure X-Ray: %w", err)
 		return
@@ -30,7 +38,7 @@ func buildConfig(ctx context.Context) (config *Config, err error) {
 
 	// At this point we're not part of a Lambda request execution, so let's
 	// explicitly create a segment to represent the configuration process.
-	ctx, segment := xray.BeginSegment(ctx, "registry.config")
+	ctx, segment := xray.BeginSegment(ctx, xraySegmentName)
 	defer func() { segment.Close(err) }()
 
 	var awsConfig aws.Config
@@ -40,11 +48,18 @@ func buildConfig(ctx context.Context) (config *Config, err error) {
 		return
 	}
 
-	secretsmanager := secrets.NewHandler(awsConfig)
+	secretsHandler := secrets.NewHandler(awsConfig)
 
-	githubAPIToken, err := secretsmanager.GetValueFromEnvVar(ctx, "GITHUB_TOKEN_SECRET_ASM_NAME")
+	githubAPIToken, err := secretsHandler.GetValueFromEnvVar(ctx, "GITHUB_TOKEN_SECRET_ASM_NAME")
 	if err != nil {
 		err = fmt.Errorf("could not get GitHub API token: %w", err)
+		return
+	}
+
+	var tableName string
+	tableName = os.Getenv("PROVIDER_VERSIONS_TABLE_NAME")
+	if tableName == "" {
+		err = fmt.Errorf("PROVIDER_VERSIONS_TABLE_NAME environment variable not set")
 		return
 	}
 
@@ -58,10 +73,23 @@ func buildConfig(ctx context.Context) (config *Config, err error) {
 	config = &Config{
 		ManagedGithubClient: github.NewManagedGithubClient(githubAPIToken),
 		RawGithubv4Client:   github.NewRawGithubv4Client(githubAPIToken),
-		ProviderRedirects:   providerRedirects,
 
-		SecretsHandler: secretsmanager,
+		SecretsHandler:       secretsHandler,
+		ProviderVersionCache: providercache.NewHandler(awsConfig, tableName),
+		LambdaClient:         lambda.NewFromConfig(awsConfig),
+
+		ProviderRedirects: providerRedirects,
+	}
+	return
+}
+
+// EffectiveProviderNamespace will map namespaces for providers in situations
+// where the author (owner of the namespace) does not release artifacts as
+// GitHub Releases.
+func (c Config) EffectiveProviderNamespace(namespace string) string {
+	if redirect, ok := c.ProviderRedirects[namespace]; ok {
+		return redirect
 	}
 
-	return
+	return namespace
 }
