+++ src/lambda/api/main.go
@@ -4,14 +4,13 @@ import (
 	"context"
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
+	"github.com/opentffoundation/registry/internal/config"
 )
 
 type LambdaFunc func(context.Context, events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error)
 
 func main() {
-	ctx := context.Background()
-
-	config, err := buildConfig(ctx)
+	config, err := config.BuildConfig(context.Background(), "registry.buildconfig")
 	if err != nil {
 		panic(err)
 	}
